Somewhat of a working implementation for attack animations.
Worked through some tutorials and read some documentation and think I have a preliminary solution for adding animation to the screen. The animation does not really play how I would want it to, and it does not spawn correclty (not center of target). Despite this, the animation does spawn on screen, dynamically located based on target, and is deleted after completion. Will need some decent tweaks before it looks like a true attack animation.